+++ packages/@webex/plugin-meetings/src/breakouts/index.ts
@@ -5,11 +5,12 @@ import {WebexPlugin} from '@webex/webex-core';
 import {debounce, forEach} from 'lodash';
 import LoggerProxy from '../common/logs/logger-proxy';
 
-import {BREAKOUTS, HTTP_VERBS, MEETINGS} from '../constants';
+import {BREAKOUTS, MEETINGS, HTTP_VERBS} from '../constants';
 
 import Breakout from './breakout';
 import BreakoutCollection from './collection';
 import BreakoutRequest from './request';
+import {boServiceErrorHandler} from './utils';
 
 /**
  * @class Breakouts
@@ -34,6 +35,7 @@ const Breakouts = WebexPlugin.extend({
     url: 'string', // appears from the moment you enable breakouts
     locusUrl: 'string', // the current locus url
     breakoutServiceUrl: 'string', // the current breakout resouce url
+    groups: 'array', // appears when create breakouts
   },
 
   children: {
@@ -49,6 +51,20 @@ const Breakouts = WebexPlugin.extend({
        */
       fn() {
         return this.sessionType === BREAKOUTS.SESSION_TYPES.MAIN;
+      },
+    },
+    breakoutGroupId: {
+      deps: ['groups'],
+      /**
+       * Returns the actived group id
+       * @returns {boolean}
+       */
+      fn() {
+        if (this.groups?.length) {
+          return this.groups[0].id;
+        }
+
+        return '';
       },
     },
   },
@@ -275,9 +291,7 @@ const Breakouts = WebexPlugin.extend({
    * @returns {Promise}
    */
   broadcast(message, options) {
-    const breakoutGroupId = this.isInMainSession
-      ? this.breakouts.filter((breakout) => !breakout.isMain)[0]?.groupId
-      : this.groupId;
+    const {breakoutGroupId} = this;
     if (!breakoutGroupId) {
       throw new Error('Cannot broadcast, no breakout session found');
     }
@@ -347,6 +361,143 @@ const Breakouts = WebexPlugin.extend({
         enableBreakoutSession: enable,
       },
     });
+  },
+
+  /**
+   * Create new breakout sessions
+   * @param {object} sessions -- breakout session group
+   * @returns {Promise}
+   */
+  async create(sessions) {
+    // @ts-ignore
+    const breakInfo = await this.webex
+      .request({
+        method: HTTP_VERBS.PUT,
+        uri: this.url,
+        body: {
+          groups: [
+            {
+              sessions,
+            },
+          ],
+        },
+      })
+      .catch((error) => {
+        return Promise.reject(
+          boServiceErrorHandler(error, 'Breakouts#create --> Edit lock token mismatch')
+        );
+      });
+
+    if (breakInfo.body?.groups) {
+      this.set('groups', breakInfo.body.groups);
+    }
+
+    return Promise.resolve(breakInfo);
+  },
+
+  /**
+   * Delete all breakout sessions
+   * @returns {Promise}
+   */
+  async clearSessions() {
+    // @ts-ignore
+    const breakInfo = await this.webex
+      .request({
+        method: HTTP_VERBS.PUT,
+        uri: this.url,
+        body: {
+          groups: [
+            {
+              action: BREAKOUTS.ACTION.DELETE,
+            },
+          ],
+        },
+      })
+      .catch((error) => {
+        return Promise.reject(
+          boServiceErrorHandler(error, 'Breakouts#clearSessions --> Edit lock token mismatch')
+        );
+      });
+
+    if (breakInfo.body?.groups) {
+      this.set('groups', breakInfo.body.groups);
+    }
+
+    return Promise.resolve(breakInfo);
+  },
+
+  /**
+   * Host or cohost starts breakout sessions
+   * @param {object} params
+   * @returns {Promise}
+   */
+  start(params = {}) {
+    const action = BREAKOUTS.ACTION.START;
+    const payload = {
+      id: this.breakoutGroupId,
+      action,
+      allowBackToMain: false,
+      allowToJoinLater: false,
+      ...params,
+    };
+
+    return this.request({
+      method: HTTP_VERBS.PUT,
+      uri: this.url,
+      body: {
+        groups: [payload],
+      },
+    }).catch((error) => {
+      return Promise.reject(
+        boServiceErrorHandler(error, 'Breakouts#start --> Edit lock token mismatch')
+      );
+    });
+  },
+
+  /**
+   * Host or cohost ends breakout sessions
+   * @param {object} params
+   * @returns {Promise}
+   */
+  end(params = {}) {
+    const {delayCloseTime, breakoutGroupId: id} = this;
+    const action = BREAKOUTS.ACTION.CLOSE;
+    const payload = {
+      id,
+      action,
+      delayCloseTime,
+      ...params,
+    };
+
+    return this.request({
+      method: HTTP_VERBS.PUT,
+      uri: this.url,
+      body: {
+        groups: [payload],
+      },
+    }).catch((error) => {
+      return Promise.reject(
+        boServiceErrorHandler(error, 'Breakouts#end --> Edit lock token mismatch')
+      );
+    });
+  },
+
+  /**
+   * get existed breakout sessions
+   * @param {boolean} editlock -- lock operations of the breakout sessions
+   * @returns {Promise}
+   */
+  async getBreakout(editlock) {
+    const breakout = await this.request({
+      method: HTTP_VERBS.GET,
+      uri: this.url + (editlock ? `?editlock=${editlock}` : ''),
+    });
+
+    if (breakout.body?.groups) {
+      this.set('groups', breakout.body.groups);
+    }
+
+    return breakout;
   },
 });